Global Styles: Add back the global styles logic that forces the solid border when color or width applied. (#50498)
